Randomise dice reset
At present, the selected dice value resets to 5 after each input of a dice roll. 

To minimise the number of user inputs during seed generation via this method, the default value should randomly oscillate between 2 and 5. From the default 5 location, two inputs are required to reach 1/3 (but only one input to reach 4/6), whereas from a default 2 location two inputs are required to reach 4/6 (but only one to 1/3). By randomly selecting either 2/5 default with 50/50 probability for either selection, statistically the number of user inputs required will be minimised during the process of seed generation by dice rolls.

This PR proposes a very simple method to randomly select a 2/5 default, based on the value of the previous input. This method is sufficiently random to benefit the user provided that the assumption that the dice used to generate the seed is unbiased, is accurate.